reservation/registrationを削除
